Refactor MIP module: reorganize imports, remove deprecated model files, and introduce new variable and solution classes
