+++ ocean/feature/__init__.py
@@ -1,11 +1,9 @@
 from .feature import Feature
 from .mapper import FeatureMapper
 from .parse import parse_features
-from .variable import FeatureVar
 
 __all__ = [
     "Feature",
     "FeatureMapper",
-    "FeatureVar",
     "parse_features",
 ]
