+++ ocean/mip/variable/__init__.py
@@ -1,0 +1,4 @@
+from .feature import FeatureVar
+from .tree import TreeVar
+
+__all__ = ["FeatureVar", "TreeVar"]
