--- ocean/model/__init__.py
@@ -1,4 +1,0 @@
-from .model import Model
-from .solution import Solution
-
-__all__ = ["Model", "Solution"]
