+++ ocean/tree/__init__.py
@@ -1,6 +1,5 @@
 from .node import Node
 from .parse import parse_tree
 from .tree import Tree
-from .variable import TreeVar
 
-__all__ = ["Node", "Tree", "TreeVar", "parse_tree"]
+__all__ = ["Node", "Tree", "parse_tree"]
