+++ tests/feature/test_variable.py
@@ -1,8 +1,9 @@
 import numpy as np
 import pytest
 
-from ocean.base import BaseModel
-from ocean.feature import Feature, FeatureVar
+from ocean.feature import Feature
+from ocean.mip import FeatureVar
+from ocean.mip.base import BaseModel
 
 from ..utils import ENV
